+ InlineHelper.dll - Ternary operator (?<-)
Note:
F# 3.0 fixed a bug in the parser that prevented to write more than two static constraints.
In order to generate 3 constraints the ternary (?<-) operator trick was used.
Now, having the relased version of F# 3.0, a .dll was compiled with the Inline Class which will generate those constraints automatically and call a static member called 'instance'.